Split in 2 to remove tail only visible at night
